ES6 compatible
Edited the script so it uses ES6 syntax and methods. The saved data is not longer attached to the element as `data` to avoid memory leaks, it now resides inside a weakmap by a reference to the root element in the chain. If the root element is removed then so is all the data associated with it, just like `data` but without all the recursive loops if retrieving the `data` from the root element.